Remove long parsing, correct interfaces

+++ Octokit.GraphQL/Model/ProjectV2FieldCommon.cs
@@ -27,7 +27,7 @@ namespace Octokit.GraphQL.Model
         /// <summary>
         /// Identifies the primary key from the database.
         /// </summary>
-        int? DatabaseId { get; }
+        long? DatabaseId { get; }
 
         /// <summary>
         /// The Node ID of the ProjectV2FieldCommon object
