+++ Octokit.GraphQL/Model/ProjectV2ItemFieldValueCommon.cs
@@ -27,7 +27,7 @@ namespace Octokit.GraphQL.Model
         /// <summary>
         /// Identifies the primary key from the database.
         /// </summary>
-        int? DatabaseId { get; }
+        long? DatabaseId { get; }
 
         /// <summary>
         /// The project field that contains this value.
